pyproject.toml
[project]
name = "heart-disease-risk-predictor"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "torch",
    "torchvision",
    "torchaudio",
    "numpy>=2.4.2",
    "pandas>=2.3.3",
    "scikit-learn>=1.8.0",
    "mlflow>=3.10.1",
    "dvc>=3.66.1",
    "pyyaml>=6.0.3",
    "matplotlib>=3.10.8",
    "seaborn>=0.13.2",
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cpu" }]
torchvision = [{ index = "pytorch-cpu" }]
torchaudio = [{ index = "pytorch-cpu" }]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "jupyter>=1.1.1",
    "pytest>=9.0.2",
]

